Playtest round 4: works-ledger crash fix, dialogue darbar court, encyclopedia akhbaar, founding coronation
Four findings from the first live round on the J-W wave:

1. CRASH (mosque build): clicking Begin in the works ledger threw inside Gauntlet's invoke —
   VillageWorksVM.BarWidth was an int bound to the widget's float SuggestedWidth, and Gauntlet
   writes layout bindings BACK into the VM; set_BarWidth(Single) vs int broke ExecuteAct and
   took the game down natively. BarWidth and CouncilEntryVM.IndentWidth are now float; the trap
   is recorded as Modding-Findings ch.19 (diagnose via "inside invoke" lines in rgl_log).

2. Darbar court reworked as DIALOGUE (was "too shallow", an inquiry screen): the sitting now
   runs as chained conversations before the throne — the plaintiff states his case and can be
   pressed for proof, the defendant stands forward and can be made to swear, the council's
   wazir/diwan leans close with political counsel (weighted by whom HE favours), and the
   sovereign speaks the judgment as a dialogue line. Chain driven by a tick pump (same pattern
   as RoyalFarmaan.Pump); ESC adjourns cleanly; sitting state is transient, never saved.
   Outcome math unchanged (DarbarCourtMath, CourtRuling records).

3. Akhbaar scouts from the encyclopedia (hunting the court-menu list was tedious): a
   "Dispatch a scout (N dinars)" button on the lord's page (UIExtenderEx mixin + prefab
   insert, farmaan button brushes), dispatchable from the map (scout rides from your camp);
   an "Akhbaar" stats row shows the scout on the road and the lord's LAST REPORTED
   whereabouts with age, persisted in the registers (hind_akhbaar_reg*).

4. Coronation on FOUNDING a kingdom (reported missing): a kingdom first sighted mid-campaign
   is a genuine accession and now stages the darbar; hind_rebel_* claim kingdoms excluded.

Also recorded: the player now runs WITH Diplomacy (reverses the 2026-07-10 decision).
Docs: PLAYTEST N4/O8/Q1/U + status ledger; wiki ch.28 rows + Modding-Findings ch.19.
340 tests green.

+++ src/TheHindostanMod/TheHindostanMod/UI/CouncilEntryVM.cs
@@ -11,7 +11,7 @@ namespace TakhtyaTaboot.UI
         private string _name;
         private string _subtitle;
         private string _detail;
-        private int _indentWidth;
+        private float _indentWidth;
         private bool _isClickable;
 
         public CouncilEntryVM(string name, string subtitle, string detail, bool isClickable, Action onActivate)
@@ -36,8 +36,11 @@ namespace TakhtyaTaboot.UI
         [DataSourceProperty]
         public bool HasDetail => !string.IsNullOrEmpty(_detail);
 
+        // FLOAT, not int: bound to a widget's SuggestedWidth (a float), and Gauntlet writes the
+        // value back into the VM — set_IndentWidth(Single) against an int throws inside invoke
+        // (the same trap that crashed the works ledger; see VillageWorksVM.BarWidth).
         [DataSourceProperty]
-        public int IndentWidth { get => _indentWidth; set { if (_indentWidth != value) { _indentWidth = value; OnPropertyChangedWithValue(value); } } }
+        public float IndentWidth { get => _indentWidth; set { if (_indentWidth != value) { _indentWidth = value; OnPropertyChangedWithValue(value); } } }
 
         [DataSourceProperty]
         public bool IsClickable { get => _isClickable; set { if (_isClickable != value) { _isClickable = value; OnPropertyChangedWithValue(value); } } }
